server: move functional.ts from types/ to utils/
