--- language-server/src/types/functional.ts
@@ -1,1 +1,0 @@
-export type Result<T, E> = [T, null] | [null, E]
