+++ language-server/src/utils/functional.ts
@@ -1,0 +1,2 @@
+export type Result<T, E> = [T, null] | [null, E]
+export const $IsError = <T, E>(r: Result<T, E>) => r[1] !== null
